* develop: (190 commits)
  Add 0.9.7 binaries for Mac OS X
  Add 0.9.7 binaries for Linux
  Add 0.9.7 binaries for Windows
  Version 0.9.7
  Fix 64-bit host detection in Windows
  Define BIG/LITTLE_ENDIAN on Linux
  Simplify endian tests, implement on Windows
  fileReadLines can cope with windows newlines
  Add warning about fileRemove behavior on Windows
  All file tests passing on Windows
  Fix crash if current exe path cannot be derived
  CMake generates nicer projects in Mac OSX
  Nicer formatting of parameters with verbose display
  Show new parameter value after setting it
  Show possible display values for parameter
  Whoops fix dumb bug in isLogLevelAtLeast
  Removed unused header
  Add headers to cmake source
  Ignore more stuff generated by cmake
  Clean up preprocessor code in test runner
  ...

+++ source/io/SampleSourceAiff.c
@@ -96,7 +96,7 @@ static boolByte _openSampleSourceAiff(void *sampleSourcePtr, const SampleSourceO
 static boolByte _readBlockFromAiffFile(void* sampleSourcePtr, SampleBuffer sampleBuffer) {
   SampleSource sampleSource = (SampleSource)sampleSourcePtr;
   SampleSourcePcmData extraData = (SampleSourcePcmData)(sampleSource->extraData);
-  int originalBlocksize = sampleBuffer->blocksize;
+  unsigned long originalBlocksize = sampleBuffer->blocksize;
   size_t samplesRead = sampleSourcePcmRead(extraData, sampleBuffer);
   sampleSource->numSamplesProcessed += samplesRead;
   return (originalBlocksize == sampleBuffer->blocksize);
@@ -105,7 +105,7 @@ static boolByte _readBlockFromAiffFile(void* sampleSourcePtr, SampleBuffer sampl
 static boolByte _writeBlockToAiffFile(void* sampleSourcePtr, const SampleBuffer sampleBuffer) {
   SampleSource sampleSource = (SampleSource)sampleSourcePtr;
   SampleSourcePcmData extraData = (SampleSourcePcmData)(sampleSource->extraData);
-  int samplesWritten = (int)sampleSourcePcmWrite(extraData, sampleBuffer);
+  unsigned int samplesWritten = (int)sampleSourcePcmWrite(extraData, sampleBuffer);
   sampleSource->numSamplesProcessed += samplesWritten;
   return (samplesWritten == sampleBuffer->blocksize);
 }
